move resources code into resources directory
